+++ lightly_studio_view/src/lib/schema.d.ts
@@ -2722,6 +2722,8 @@ export interface components {
             collection_id?: string | null;
             /** Annotation Label Ids */
             annotation_label_ids?: string[] | null;
+            /** Include No Annotations */
+            include_no_annotations?: boolean | null;
             /** Tag Ids */
             tag_ids?: string[] | null;
             /** Metadata Filters */
@@ -3055,6 +3057,8 @@ export interface components {
             video_filter?: components["schemas"]["VideoFilter"] | null;
             /** Video Frames Annotations Labels */
             video_frames_annotations_labels?: string[] | null;
+            /** Include No Annotations */
+            include_no_annotations?: boolean | null;
         };
         /**
          * VideoFieldsBoundsBody
@@ -3085,6 +3089,8 @@ export interface components {
             duration_s?: components["schemas"]["FloatRange"] | null;
             /** Annotation Frames Label Ids */
             annotation_frames_label_ids?: string[] | null;
+            /** Include No Annotations */
+            include_no_annotations?: boolean | null;
             sample_filter?: components["schemas"]["SampleFilter"] | null;
         };
         /**
@@ -3124,6 +3130,8 @@ export interface components {
             video_filter?: components["schemas"]["VideoFrameFilter"] | null;
             /** Annotations Labels */
             annotations_labels?: string[] | null;
+            /** Include No Annotations */
+            include_no_annotations?: boolean | null;
         };
         /**
          * VideoFrameFieldsBoundsView
@@ -4397,6 +4405,7 @@ export interface operations {
         parameters: {
             query?: {
                 filtered_labels?: string[] | null;
+                include_no_annotations?: boolean | null;
                 min_width?: number | null;
                 max_width?: number | null;
                 min_height?: number | null;